+++ test/Main.hs
@@ -50,9 +50,7 @@ import NovaEngine.Mesh.Types
 import NovaEngine.Mesh.UV
 import NovaEngine.Mesh.Weld
 import NovaEngine.Noise
-import NovaEngine.Render.Material
-import NovaEngine.Render.Skin (SkinnedVertex (..))
-import NovaEngine.Render.Texture (calcMipLevels)
+import NovaEngine.Render.Types
 import NovaEngine.SDF
 import NovaEngine.SDF.DualContour
 import NovaEngine.SDF.Isosurface